Make SingleElementStreamWriter top-level
