+++ events/src/main/java/org/opentest4j/reporting/events/api/DefaultDocumentWriter.java
@@ -19,7 +19,6 @@ package org.opentest4j.reporting.events.api;
 import org.opentest4j.reporting.schema.Namespace;
 import org.opentest4j.reporting.schema.QualifiedName;
 
-import javax.xml.namespace.NamespaceContext;
 import javax.xml.stream.XMLOutputFactory;
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
@@ -37,7 +36,6 @@ import java.util.function.Consumer;
 import static java.util.stream.Collectors.joining;
 import static javax.xml.XMLConstants.FEATURE_SECURE_PROCESSING;
 import static javax.xml.XMLConstants.W3C_XML_SCHEMA_INSTANCE_NS_URI;
-import static org.opentest4j.reporting.events.api.Element.CDATA_WORKAROUND_COMMENT;
 import static org.opentest4j.reporting.events.api.NamespaceRegistry.XSI_PREFIX;
 
 class DefaultDocumentWriter<T extends Element<T>> implements DocumentWriter<T> {
@@ -142,175 +140,4 @@ class DefaultDocumentWriter<T extends Element<T>> implements DocumentWriter<T> {
 		bufferedWriter.flush();
 	}
 
-	private static class SingleElementStreamWriter implements XMLStreamWriter {
-
-		private final XMLStreamWriter delegate;
-		private final NamespaceRegistry namespaceRegistry;
-
-		public SingleElementStreamWriter(XMLStreamWriter delegate, NamespaceRegistry namespaceRegistry) {
-			this.delegate = delegate;
-			this.namespaceRegistry = namespaceRegistry;
-		}
-
-		@Override
-		public void writeStartElement(String localName) throws XMLStreamException {
-			delegate.writeStartElement(localName);
-		}
-
-		@Override
-		public void writeStartElement(String namespaceURI, String localName) throws XMLStreamException {
-			delegate.writeStartElement(namespaceURI, localName);
-		}
-
-		@Override
-		public void writeStartElement(String prefix, String localName, String namespaceURI) throws XMLStreamException {
-			delegate.writeStartElement(prefix, localName, namespaceURI);
-		}
-
-		@Override
-		public void writeEmptyElement(String namespaceURI, String localName) throws XMLStreamException {
-			delegate.writeEmptyElement(namespaceURI, localName);
-		}
-
-		@Override
-		public void writeEmptyElement(String prefix, String localName, String namespaceURI) throws XMLStreamException {
-			delegate.writeEmptyElement(prefix, localName, namespaceURI);
-		}
-
-		@Override
-		public void writeEmptyElement(String localName) throws XMLStreamException {
-			delegate.writeEmptyElement(localName);
-		}
-
-		@Override
-		public void writeEndElement() throws XMLStreamException {
-			delegate.writeEndElement();
-		}
-
-		@Override
-		public void writeEndDocument() {
-		}
-
-		@Override
-		public void close() throws XMLStreamException {
-			delegate.close();
-		}
-
-		@Override
-		public void flush() throws XMLStreamException {
-			delegate.flush();
-		}
-
-		@Override
-		public void writeAttribute(String localName, String value) throws XMLStreamException {
-			delegate.writeAttribute(localName, value);
-		}
-
-		@Override
-		public void writeAttribute(String prefix, String namespaceURI, String localName, String value)
-				throws XMLStreamException {
-			delegate.writeAttribute(prefix, namespaceURI, localName, value);
-		}
-
-		@Override
-		public void writeAttribute(String namespaceURI, String localName, String value) throws XMLStreamException {
-			delegate.writeAttribute(namespaceURI, localName, value);
-		}
-
-		@Override
-		public void writeNamespace(String prefix, String namespaceURI) throws XMLStreamException {
-			if (!namespaceRegistry.containsUri(namespaceURI)) {
-				delegate.writeNamespace(prefix, namespaceURI);
-			}
-		}
-
-		@Override
-		public void writeDefaultNamespace(String namespaceURI) {
-		}
-
-		@Override
-		public void writeComment(String data) throws XMLStreamException {
-			// Remove bogus comment that avoids CDATA sections from being merged
-			if (!CDATA_WORKAROUND_COMMENT.equals(data)) {
-				delegate.writeComment(data);
-			}
-		}
-
-		@Override
-		public void writeProcessingInstruction(String target) throws XMLStreamException {
-			delegate.writeProcessingInstruction(target);
-		}
-
-		@Override
-		public void writeProcessingInstruction(String target, String data) throws XMLStreamException {
-			delegate.writeProcessingInstruction(target, data);
-		}
-
-		@Override
-		public void writeCData(String data) throws XMLStreamException {
-			delegate.writeCData(data);
-		}
-
-		@Override
-		public void writeDTD(String dtd) throws XMLStreamException {
-			delegate.writeDTD(dtd);
-		}
-
-		@Override
-		public void writeEntityRef(String name) throws XMLStreamException {
-			delegate.writeEntityRef(name);
-		}
-
-		@Override
-		public void writeStartDocument() {
-		}
-
-		@Override
-		public void writeStartDocument(String version) {
-		}
-
-		@Override
-		public void writeStartDocument(String encoding, String version) {
-		}
-
-		@Override
-		public void writeCharacters(String text) throws XMLStreamException {
-			delegate.writeCharacters(text);
-		}
-
-		@Override
-		public void writeCharacters(char[] text, int start, int len) throws XMLStreamException {
-			delegate.writeCharacters(text, start, len);
-		}
-
-		@Override
-		public String getPrefix(String uri) throws XMLStreamException {
-			return delegate.getPrefix(uri);
-		}
-
-		@Override
-		public void setPrefix(String prefix, String uri) throws XMLStreamException {
-			delegate.setPrefix(prefix, uri);
-		}
-
-		@Override
-		public void setDefaultNamespace(String uri) throws XMLStreamException {
-			delegate.setDefaultNamespace(uri);
-		}
-
-		@Override
-		public void setNamespaceContext(NamespaceContext context) throws XMLStreamException {
-			delegate.setNamespaceContext(context);
-		}
-
-		@Override
-		public NamespaceContext getNamespaceContext() {
-			return delegate.getNamespaceContext();
-		}
-
-		@Override
-		public Object getProperty(String name) throws IllegalArgumentException {
-			return delegate.getProperty(name);
-		}
-	}
 }
